third_party/WebKit/LayoutTests/fast/css3-text/css3-text-decoration/getComputedStyle/getComputedStyle-text-underline-position-expected.txt

Test to make sure text-underline-position is computed properly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Initial value:
PASS e.style.textUnderlinePosition is ""
PASS computedStyle.textUnderlinePosition is "auto"


Value '':
PASS e.style.textUnderlinePosition is ""
PASS computedStyle.textUnderlinePosition is "auto"


Initial value (explicit):
PASS e.style.textUnderlinePosition is "initial"
PASS computedStyle.textUnderlinePosition is "auto"


Value 'auto':
PASS e.style.textUnderlinePosition is "auto"
PASS computedStyle.textUnderlinePosition is "auto"


Value 'under':
PASS e.style.textUnderlinePosition is "under"
PASS computedStyle.textUnderlinePosition is "under"


Value 'left':
PASS e.style.textUnderlinePosition is "left"
PASS computedStyle.textUnderlinePosition is "left"


Value 'right':
PASS e.style.textUnderlinePosition is "right"
PASS computedStyle.textUnderlinePosition is "right"


Value 'under left':
PASS e.style.textUnderlinePosition is "under left"
PASS computedStyle.textUnderlinePosition is "under left"


Value 'right under':
PASS e.style.textUnderlinePosition is "under right"
PASS computedStyle.textUnderlinePosition is "under right"


Ancestor inherits values from parent:
PASS e.style.textUnderlinePosition is ""
PASS computedStyle.textUnderlinePosition is "under"


Value 'auto under':
PASS e.style.textUnderlinePosition is ""
PASS computedStyle.textUnderlinePosition is "under"


Value 'auto left':
PASS e.style.textUnderlinePosition is ""
PASS computedStyle.textUnderlinePosition is "under"


Value 'right auto':
PASS e.style.textUnderlinePosition is ""
PASS computedStyle.textUnderlinePosition is "under"


Value 'under under':
PASS e.style.textUnderlinePosition is ""
PASS computedStyle.textUnderlinePosition is "under"


Value 'left right':
PASS e.style.textUnderlinePosition is ""
PASS computedStyle.textUnderlinePosition is "under"


Value 'auto alphabetic under':
PASS e.style.textUnderlinePosition is ""
PASS computedStyle.textUnderlinePosition is "under"


PASS successfullyParsed is true

TEST COMPLETE
